update pw_mcjax_dag, with imagee, kpo and pass workload_id
