+++ dags/maxtext_pathways/configs/parameters.py
@@ -17,6 +17,7 @@
 from airflow.models.param import Param
 from dags.maxtext_pathways.configs import model_configs as model_cfg
 from dags.common.vm_resource import TpuVersion
+from dags.common.vm_resource import DockerImage
 
 MODEL_FRAMEWORK = ["mcjax", "pathways"]
 
@@ -31,6 +32,8 @@ MODEL_NAME.extend(V5P_MODEL_NAME)
 MODEL_NAME.extend(V6E_MODEL_NAME)
 
 DEVICE_VERSION = ["v" + version.value for version in TpuVersion]
+
+IMAGE = DockerImage.MAXTEXT_TPU_JAX_STABLE
 
 PARAMETERS = {
     "user": Param(
@@ -92,22 +95,21 @@ PARAMETERS = {
         description="Number of slices",
     ),
     "server_image": Param(
-        # TODO(b/451750407): Replace this temporary image with a formal one.
-        "gcr.io/cienet-cmcs/lidanny/unsanitized_server:latest",
+        # Reference to https://g3doc.corp.google.com/cloud/tpu/g3doc/fas/pathways-on-cloud/index.md?cl=head
+        "us-docker.pkg.dev/cloud-tpu-v2-images/pathways/server:20260102-jax_0.8.1",
         type="string",
         title="Server Image",
         description="Server image for pathways.",
     ),
     "proxy_image": Param(
-        # TODO(b/451750407): Replace this temporary image with a formal one.
-        "gcr.io/cienet-cmcs/lidanny/unsanitized_proxy_server:latest",
+        # Reference to https://g3doc.corp.google.com/cloud/tpu/g3doc/fas/pathways-on-cloud/index.md?cl=head
+        "us-docker.pkg.dev/cloud-tpu-v2-images/pathways/proxy_server:20260102-jax_0.8.1",
         type="string",
         title="Proxy Image",
         description="Proxy image for pathways.",
     ),
     "runner": Param(
-        # TODO(b/451750407): Replace this temporary image with a formal one.
-        "gcr.io/cienet-cmcs/lidanny_latest:latest",
+        IMAGE.value,
         type="string",
         title="Runner Image",
         description="Runner image for the cluster.",
